Move tests to CommandersActStreamingAnalyticsStateTransitionTests
